Make a separate class for TextWrapper
